# src/api/dtos/base_dto.go
package dtos

type CreateUpdateCountryRequest struct {
	Name string `json:"name" binding:"required,alpha,min=3,max=20"`
}

type CountryResponse struct {
	Id   int    `json:"id"`
	Name string `json:"name"`
}